agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nexus"
version = "0.2.0"
requires-python = ">=3.12,<3.14"
dependencies = [
    "uvicorn[standard]>=0.29",
    "openai>=1.50",
    "anthropic>=0.40",
    "tomli_w",
    "psutil>=5.9",
    "python-multipart>=0.0.26",
    "fastembed>=0.4",
    "spacy>=3.8",
    "langdetect>=1.0",
    # Bundle the spaCy en_core_web_sm and pt_core_news_sm models so the
    # builtin extractor works offline on first run — pinned because spaCy
    # model wheels are version-locked to spaCy minor versions.
    "en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl",
    "pt-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.8.0/pt_core_news_sm-3.8.0-py3-none-any.whl",
    "faster-whisper>=1.0",
    "loom-framework[anthropic,acp,tui,graphrag,search,scrape]",
    "huggingface_hub>=0.24",
    "duckdb>=1.0",
    "python-dateutil>=2.9",
    "pywebpush>=2.0",
    "cryptography>=42.0",
    # qrcode renders a scannable code in the CLI's `tunnel start`. The tunnel
    # itself uses Cloudflare Quick Tunnel via the `cloudflared` binary, which
    # we auto-download on first use (see nexus.tunnel.cloudflared_provider).
    "qrcode>=7.4",
    # Tiny pure-Python BM25 implementation. Used by the bundled `nexus` skill
    # (`nexus_kb_search` tool) to retrieve sections from the in-app knowledge
    # base without inflating the system prompt.
    "rank_bm25>=0.2.2",
]

[tool.uv.sources]
# Loom (the agentic-core framework) is pulled straight from GitHub so the
# one-line installer works on a fresh host. For local development against a
# sibling checkout, override this with a `[tool.uv.sources]` block in a
# user-level `uv.toml` (or `uv pip install -e ../../loom` after `uv sync`).
loom-framework = { git = "https://github.com/NinoCoelho/loom.git", branch = "main" }
ddgs = { git = "https://github.com/NinoCoelho/ddgs", branch = "main" }

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
    "pytest-asyncio",
]
# Bundled `pdf-maker` skill needs Pillow + fpdf2 to render PDFs.
# Install with: `uv sync --extra pdf`
pdf = [
    "Pillow>=10.0",
    "fpdf2>=2.7",
]
# AWS Bedrock support. Install with: `uv sync --extra bedrock` (or
# `uv tool install nexus[bedrock]`). Pulls in boto3 for SigV4 signing
# + the InvokeModel/Converse APIs. The catalog tile is visible without
# this extra; the wizard's connection test surfaces a clear "install
# nexus[bedrock]" hint when boto3 isn't importable.
bedrock = [
    "boto3>=1.34",
]
# Bundled `markitdown` skill converts arbitrary docs (PDF/DOCX/PPTX/XLSX/…) to
# Markdown so the agent and the vault FTS index can read them. The wrapper
# script also pip-installs on first run if this extra wasn't preinstalled.
# Install with: `uv sync --extra markitdown`
markitdown = [
    "markitdown[all]>=0.0.1",
]

[project.scripts]
nexus = "nexus.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/nexus"]

[tool.hatch.metadata]
# Required so direct URL deps (e.g. the en_core_web_sm wheel) are honored.
allow-direct-references = true

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
# Treat async def tests as asyncio by default so we don't have to
# sprinkle ``@pytest.mark.asyncio`` on every HITL test.
asyncio_mode = "auto"
markers = [
    "slow: tests that download multi-MB models or hit the network",
]
